Add files via upload
Changes to php file. Both html and php code are now combined in single page. Please note the following variables may need to be updated depending on your database and settings:

$user                             // user name was 'root' in my case
$password                    // leave blank if you don't have any password set to access the DB
$db                               // name of your database

Changes:

The page now displays result of your query in form of a table with the schema  of the table in green color as  well as every even and odd row having a different color.